+++ CopilotKit/packages/cloud/src/types/service-adapter.ts
@@ -3,10 +3,3 @@ import { AnnotatedFunction } from "@copilotkit/shared";
 export interface CopilotKitServiceAdapter {
   stream(functions: AnnotatedFunction<any[]>[], forwardedProps: any): ReadableStream;
 }
-
-export type CopilotKitOpenAIConfiguration = {
-  provider?: "openai"; // default to openai
-  apiKey?: string;
-  model?: string;
-  functions?: AnnotatedFunction<any[]>[];
-};
