+++ CopilotKit/packages/cloud/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./utils";
 export * from "./openai";
+export * from "./openai-stream-interceptor";
